# webApp/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/minecraft/versions": {
            "get": {
                "description": "Returns list of available Minecraft versions and server cores",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minecraft"
                ],
                "summary": "Get available Minecraft versions and cores",
                "responses": {
                    "200": {
                        "description": "Returns versions and cores",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/server/restart": {
            "post": {
                "description": "Перезагружает сервер Minecraft через docker-compose",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minecraft"
                ],
                "summary": "Перезагрузка сервера Minecraft",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/server/start": {
            "post": {
                "description": "Запускает сервер Minecraft через docker-compose",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minecraft"
                ],
                "summary": "Запуск сервера Minecraft",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/server/stop": {
            "post": {
                "description": "Останавливает сервер Minecraft через docker-compose",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minecraft"
                ],
                "summary": "Остановка сервера Minecraft",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}